% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder2.R
\docType{data}
\name{gapminder2}
\alias{gapminder2}
\title{gapminder2}
\format{A dataframe with 1064 observation and 7 variables
\describe{
\item{country}{countries in dataset}
\item{contienent}{continent countries belong to}
\item{year}{year for the observation variables}
\item{lifeExp}{mean life expectancy of the countries}
\item{pop}{population of countries}
\item{gdpPercap}{GDP of countries}
\item{meanSchool}{mean years spent in school for men aged between 25-34 years old}
...
}}
\source{
gapminder package
}
\usage{
gapminder2
}
\description{
gapminder2
}
\keyword{datasets}
