% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder3.R
\docType{data}
\name{gapminder3}
\alias{gapminder3}
\title{gapminder3}
\format{object of class tibble inherited from class table and dataframe}
\source{
gapminder2 package
}
\usage{
gapminder3
}
\description{
gapminder3
}
\keyword{datasets}
