% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminderplus-package.R
\docType{package}
\name{gapminderplus}
\alias{gapminderplus}
\alias{gapminderplus-package}
\title{gapminderplus: A data package and an extention of the existing gapminder dataset with subsetting by country capabilities}
\description{
Gapminder2, and subsetting capabilities
Gapminder3, and subsetting capabilities
}
\note{
Future package development will provide an inclusion of further subsets and plotting options
}
\author{
Seevasant Indran
}
\keyword{gapminder2}
\keyword{internal}
