% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_country.R
\name{subset_country}
\alias{subset_country}
\alias{subset_country2}
\title{Subset a country using a set of inputs via the console}
\usage{
subset_country()

subset_country2()
}
\arguments{
\item{Provide}{the country or countries without " " in the console after executing \code{subset_country()}.}
}
\value{
A list of country that is subseted from the gapminder2 dataset for subset_country

A list of country that is subseted from the gapminder2 dataset for subset_country2
}
\description{
Type the country or countries in the console to subset the gapminder dataset more interactively.
}
\details{
Provide a list in the console to subset the gapminder dataset more interactively.
}
\examples{
#' subset_country()
#' subset_country2()
}
