% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder2.R, R/gapminderplus-package.R
\docType{data}
\name{gapminder2}
\alias{gapminder2}
\alias{gapminderplus}
\alias{gapminderplus-package}
\alias{gapminder2-package}
\title{gapminder2}
\format{object of class tibble inherited from class table and dataframe}
\source{
gapminder package
}
\usage{
gapminder2
}
\description{
One more column (mean age in school) is added onto the dataset and more on the way (check raw data), and subsetting capabilities
}
\note{
Future package development will provide an inclusion of further subsets and plotting options
}
\author{
Seevasant Indran
}
\keyword{datasets}
\keyword{gapminder2}
\keyword{internal}
