% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder3.R
\docType{data}
\name{gapminder3}
\alias{gapminder3}
\title{gapminder3}
\format{A dataframe with 524 observation and 8 variables
\describe{
\item{country}{131 countries in dataset}
\item{contienent}{continent countries belong to}
\item{year}{year for the observation variables}
\item{lifeExp}{mean life expectancy of the countries}
\item{pop}{population of countries}
\item{gdpPercap}{GDP of countries}
\item{meanSchool}{mean years spent in school for men aged between 25-34 years old}
\item{infantMortality}{infant mortality observed in each country}
...
}}
\source{
gapminderplus package
}
\usage{
gapminder3
}
\description{
gapminder3
}
\keyword{datasets}
